Niveles/4/mlops-proyecto-final/airflow/requirements.txt
apache-airflow==2.7.0
mlflow==2.8.0
psycopg2-binary==2.9.9
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
requests==2.31.0
scipy==1.11.4
boto3==1.29.7
